requirements.txt
Django==2.0.7
django-cors-headers==2.4.0
djangorestframework==3.8.2
dj-rest-auth==1.1.1
django-rest-auth== 0.9.5
django-allauth==0.42.0
pytz==2018.5

